refactor: move migrationMonitor out of live-migration-source.go
